[FEAT] Extension 정리
