+++ Extensions/UIScrollView+Extension.swift
@@ -1,0 +1,25 @@
+//
+//  UIScrollView+Extension.swift
+//  BookMark
+//
+//  Created by JOSUEYEON on 2023/03/23.
+//
+
+import UIKit
+
+// MARK: - UIScrollView extension: 동적 크기 계산용
+extension UIScrollView {
+    func updateContentSize() {
+        let unionCalculatedTotalRect = recursiveUnionInDepthFor(view: self)
+        self.contentSize = CGSize(width: self.frame.width, height: unionCalculatedTotalRect.height+50)
+    }
+    
+    private func recursiveUnionInDepthFor(view: UIView) -> CGRect {
+        var totalRect: CGRect = .zero
+    
+        for subView in view.subviews {
+            totalRect = totalRect.union(recursiveUnionInDepthFor(view: subView))
+        }
+        return totalRect.union(view.frame)
+    }
+}
