+++ Extensions/UIView+Extension.swift
@@ -1,0 +1,15 @@
+//
+//  UIView+Extension.swift
+//  BookMark
+//
+//  Created by JOSUEYEON on 2023/03/23.
+//
+
+import UIKit
+
+// MARK: - view extension
+extension UIView {
+    func addSubviews(_ views: UIView...) {
+        views.forEach { addSubview($0) }
+    }
+}
