+++ View/MyLib/NetworkFailMainView.swift
@@ -1,0 +1,8 @@
+//
+//  NetworkFailMainView.swift
+//  BookMark
+//
+//  Created by JOSUEYEON on 2023/03/23.
+//
+
+import Foundation
